Various setjmp/longjmp fixes (Closes: #198)
Firstly, (sig)jmp_buf's contents is turned into a union on CHERI to
ensure the correct alignment, that tag bits are propagated when copied
around (needed for csh) and that capabilities are always at the right
aligned location within the buffer even when copied to a different
buffer.

In the process, the broken CHERI code for the hybrid versions was fixed
to use a0 for the incoming address rather than the garbage $c3, and the
purecap longjmp gained a missing DDC restore.